Fixed camera inside light sphere problem by using additional stencil buffer pass.

+++ samples/Common/ISample.h
@@ -16,7 +16,7 @@ public:
 	virtual bool Init(gls::CreateContextInfo& info) = 0;
 	virtual void Deinit() = 0;
 	virtual void GetFramebufferFormat(gls::FramebufferFormat& fbufFormat) = 0;
-	virtual void Render(int frame_time) = 0;
+	virtual void Render(float frame_time) = 0;
 	virtual void OnResize(int width, int height) = 0;
 	virtual void OnKeyDown(int key)				{ }
 	virtual void OnKeyUp(int key)				{ }
